+++ src/librustc/middle/trans/reflect.rs
@@ -309,7 +309,7 @@ pub impl Reflector {
             do self.bracketed(~"enum", enum_args) |this| {
                 for variants.eachi |i, v| {
                     let variant_args = ~[this.c_uint(i),
-                                         this.c_int(v.disr_val),
+                                         this.c_int(v.disr_val /*bad*/as int),
                                          this.c_uint(v.args.len()),
                                          this.c_slice(ccx.sess.str_of(v.name))];
                     do this.bracketed(~"enum_variant", variant_args) |this| {
